nearest_neighbor_analysis: Generate nearest neighbor comparisons for isolate pairs

This function will compare the isolate to their nearest neighbor. Has functionality for categorical and continuous variables.

Usage

nearest_neighbor_analysis(
  comparison_df,
  categorical_vars = NULL,
  binary_vars = NULL,
  continuous_vars = NULL,
  log_2 = FALSE
)

Arguments

comparison_df

Dataframe with isolates and their nearest neighbors.

categorical_vars

Categorical variables. Must be in the comparison_df

binary_vars

Binary variables. Must be in the comparison_df AND coded 0 and 1.

continuous_vars

Continuous variables. Must be in the comparison_df

log_2

Boolean (TRUE/FALSE) Whether to perform log-2 transformation on continuous variables

Value

Dataframe with differences in the categorical and continuous variables between an isolate and their nearest neighbor

Details

This function will use the phylogenetic tree to identify an isolate's nearest neighbor.